// src/DockiUp.Frontend/src/app/api/model/notificationChannelDto.ts
/**
 * DockiUp.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { NotificationChannelType } from './notificationChannelType';


export interface NotificationChannelDto { 
    id: string;
    name: string;
    type: NotificationChannelType;
    enabled: boolean;
    events: Array<string>;
    url: string | null;
    target: string | null;
    headerName: string | null;
    secretMasked: string | null;
    createdAt: string;
}
export namespace NotificationChannelDto {
}