// src/DockiUp.Frontend/src/app/api/model/notificationChannelType.ts
/**
 * DockiUp.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type NotificationChannelType = 'Discord' | 'Telegram' | 'Ntfy' | 'Slack' | 'Webhook';

export const NotificationChannelType = {

    Discord: 'Discord' as NotificationChannelType,

    Telegram: 'Telegram' as NotificationChannelType,

    Ntfy: 'Ntfy' as NotificationChannelType,

    Slack: 'Slack' as NotificationChannelType,

    Webhook: 'Webhook' as NotificationChannelType
};