// src/DockiUp.Frontend/src/app/api/model/saveNotificationChannelRequest.ts
/**
 * DockiUp.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { NotificationChannelType } from './notificationChannelType';


export interface SaveNotificationChannelRequest { 
    name: string;
    type: NotificationChannelType;
    enabled: boolean;
    events: Array<string>;
    url: string | null;
    target: string | null;
    headerName: string | null;
    secret: string | null;
}
export namespace SaveNotificationChannelRequest {
}